+++ CHANGELOG.md
@@ -26,6 +26,8 @@ Changes:
 - `Captions.save` output_format is now case insensitive
 - Improved MicroDVD style conversion
 - Internal restructure for faster development
+- Invalid `style` argument will result in `style=None`
+- Added `style_options` for changing style globaly, default `style="full"` `lines=-1`
 
 Fixes:
 - Fixed `Styling.getTTML` converting invalid css properties into ttml properties. To-do: add value checks for these properties.